Ranju/Shruti - fixed bugs in passing.
What changes have you made?
changes are been made on the public-javascript and controller to show
the pass button after completion of each round.
*

Why do we need this change?
to make the 'pass' functionality works properly.
*

What can this change impact?
Now the game can be played according to the rules.
*